+++ science/a_scie.py
@@ -11,75 +11,78 @@ from pathlib import Path
 
 from packaging.version import Version
 
-from science.fetcher import fetch_and_verify
+from science.fetcher import FetchResult, fetch_and_verify
 from science.hashing import Digest, Fingerprint
-from science.model import File, Ptex, ScieJump, Url
-from science.platform import Platform
+from science.model import Ptex, ScieJump, Url
+from science.platform import CURRENT_PLATFORM, Platform
 
 
 @dataclass(frozen=True)
-class _LoadResult:
-    path: Path
+class LoadResult(FetchResult):
     binary_name: str
 
 
-def _load_project_release(
+def load_project_release(
     project_name: str,
     binary_name: str,
     version: Version | None = None,
     fingerprint: Digest | Fingerprint | None = None,
-    platform: Platform = Platform.current(),
-) -> _LoadResult:
+    platform: Platform = CURRENT_PLATFORM,
+    base_url: Url | None = None,
+) -> LoadResult:
     qualified_binary_name = platform.qualified_binary_name(binary_name)
-    base_url = f"https://github.com/a-scie/{project_name}/releases"
+    root_url = (base_url or f"https://github.com/a-scie/{project_name}/releases").rstrip("/")
     if version:
         version_path = f"download/v{version}"
         ttl = None
     else:
         version_path = "latest/download"
         ttl = timedelta(days=5)
-    path = fetch_and_verify(
-        url=Url(f"{base_url}/{version_path}/{qualified_binary_name}"),
+    result = fetch_and_verify(
+        url=Url(f"{root_url}/{version_path}/{qualified_binary_name}"),
         fingerprint=fingerprint,
         executable=True,
         ttl=ttl,
     )
-    return _LoadResult(path=path, binary_name=qualified_binary_name)
+    return LoadResult(path=result.path, digest=result.digest, binary_name=qualified_binary_name)
 
 
-def jump(specification: ScieJump | None = None, platform: Platform = Platform.current()) -> Path:
+def jump(
+    specification: ScieJump | None = None, platform: Platform = CURRENT_PLATFORM
+) -> LoadResult:
     version = specification.version if specification else None
     fingerprint = specification.digest if specification and specification.digest else None
-    return _load_project_release(
+    base_url = specification.base_url if specification else None
+    return load_project_release(
         project_name="jump",
         binary_name="scie-jump",
         version=version,
         fingerprint=fingerprint,
         platform=platform,
-    ).path
+        base_url=base_url,
+    )
 
 
-def custom_jump(repo_path: Path) -> Path:
+def custom_jump(repo_path: Path) -> LoadResult:
     dist_dir = tempfile.mkdtemp()
     atexit.register(shutil.rmtree, dist_dir, ignore_errors=True)
     subprocess.run(
         args=["cargo", "run", "-p", "package", "--", dist_dir], cwd=repo_path, check=True
     )
-    return Path(dist_dir) / Platform.current().qualified_binary_name("scie-jump")
+    qualified_binary_name = CURRENT_PLATFORM.qualified_binary_name("scie-jump")
+    path = Path(dist_dir) / qualified_binary_name
+    return LoadResult(path=path, digest=Digest.hash(path), binary_name=qualified_binary_name)
 
 
-def ptex(
-    dest_dir: Path, specification: Ptex | None = None, platform: Platform = Platform.current()
-) -> File:
+def ptex(specification: Ptex | None = None, platform: Platform = CURRENT_PLATFORM) -> LoadResult:
     version = specification.version if specification else None
     fingerprint = specification.digest if specification and specification.digest else None
-    result = _load_project_release(
+    base_url = specification.base_url if specification else None
+    return load_project_release(
         project_name="ptex",
-        fingerprint=fingerprint,
         binary_name="ptex",
         version=version,
+        fingerprint=fingerprint,
         platform=platform,
+        base_url=base_url,
     )
-    (dest_dir / result.binary_name).symlink_to(result.path)
-    ptex_key = specification.id if specification and specification.id else "ptex"
-    return File(name=result.binary_name, key=ptex_key, is_executable=True)